Addressing code comments: renaming of classes, variables etc, splitting ccd definitions into multiple files, adding CCD comments
