+++ ccd-definition/AuthorisationCaseField/CareSupervision/cafcass-solicitor.json
@@ -338,7 +338,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "uploadDocumentsRouter",
+    "CaseFieldID": "manageDocumentsAction",
     "UserRole": "[CAFCASSSOLICITOR]",
     "CRUD": "CRU"
   },
