+++ ccd-definition/AuthorisationCaseField/CareSupervision/court-admin.json
@@ -1124,7 +1124,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "uploadDocumentsRouter",
+    "CaseFieldID": "manageDocumentsAction",
     "UserRole": "caseworker-publiclaw-courtadmin",
     "CRUD": "CRU"
   },
