+++ ccd-definition/AuthorisationCaseField/CareSupervision/judiciary.json
@@ -1017,7 +1017,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "uploadDocumentsRouter",
+    "CaseFieldID": "manageDocumentsAction",
     "UserRole": "caseworker-publiclaw-judiciary",
     "CRUD": "CRU"
   },
