+++ ccd-definition/AuthorisationCaseField/CareSupervision/solicitor.json
@@ -214,7 +214,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "uploadDocumentsRouter",
+    "CaseFieldID": "manageDocumentsAction",
     "UserRole": "[SOLICITOR]",
     "CRUD": "CRU"
   },
