--- ccd-definition/CaseEventToComplexTypes/manage-documents/deletedCourtDocument.json
@@ -1,59 +1,0 @@
-[
-  {
-    "LiveFrom": "01/01/2017",
-    "ID": "CourtAdminDocument",
-    "CaseEventID": "uploadStandardDirections",
-    "CaseFieldID": "deletedCourtDocument",
-    "ListElementCode": "documentTitle",
-    "FieldDisplayOrder": 1,
-    "DisplayContext": "READONLY"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "ID": "CourtAdminDocument",
-    "CaseEventID": "uploadStandardDirections",
-    "CaseFieldID": "deletedCourtDocument",
-    "EventElementLabel": "File",
-    "ListElementCode": "document",
-    "FieldDisplayOrder": 2,
-    "DisplayContext": "READONLY"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "ID": "CourtAdminDocument",
-    "CaseEventID": "uploadOtherCourtAdminDocuments",
-    "CaseFieldID": "deletedCourtDocument",
-    "ListElementCode": "documentTitle",
-    "FieldDisplayOrder": 1,
-    "DisplayContext": "READONLY"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "ID": "CourtAdminDocument",
-    "CaseEventID": "uploadOtherCourtAdminDocuments",
-    "CaseFieldID": "deletedCourtDocument",
-    "EventElementLabel": "File",
-    "ListElementCode": "document",
-    "FieldDisplayOrder": 2,
-    "DisplayContext": "READONLY"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "ID": "CourtAdminDocument",
-    "CaseEventID": "limitedUploadDocuments",
-    "CaseFieldID": "deletedCourtDocument",
-    "ListElementCode": "documentTitle",
-    "FieldDisplayOrder": 1,
-    "DisplayContext": "READONLY"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "ID": "CourtAdminDocument",
-    "CaseEventID": "limitedUploadDocuments",
-    "CaseFieldID": "deletedCourtDocument",
-    "EventElementLabel": "File",
-    "ListElementCode": "document",
-    "FieldDisplayOrder": 2,
-    "DisplayContext": "READONLY"
-  }
-]
