+++ ccd-definition/CaseEventToComplexTypes/manage-documents/manage-documents.json
@@ -1,0 +1,21 @@
+[
+  {
+    "LiveFrom": "01/01/2017",
+    "ID": "CourtAdminDocument",
+    "CaseEventID": "manageDocuments",
+    "CaseFieldID": "deletedCourtDocument",
+    "ListElementCode": "documentTitle",
+    "FieldDisplayOrder": 1,
+    "DisplayContext": "READONLY"
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "ID": "CourtAdminDocument",
+    "CaseEventID": "manageDocuments",
+    "CaseFieldID": "deletedCourtDocument",
+    "EventElementLabel": "File",
+    "ListElementCode": "document",
+    "FieldDisplayOrder": 2,
+    "DisplayContext": "READONLY"
+  }
+]
