+++ ccd-definition/CaseEventToFields/CareSupervision/limitedUploadDocuments.json
@@ -3,8 +3,8 @@
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseEventID": "limitedUploadDocuments",
-    "CaseFieldID": "uploadDocumentsRouter",
-    "PageFieldDisplayOrder": 2,
+    "CaseFieldID": "manageDocumentsAction",
+    "PageFieldDisplayOrder": 1,
     "DisplayContext": "MANDATORY",
     "PageID": "routing",
     "PageDisplayOrder": 1,
@@ -17,12 +17,12 @@
     "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseEventID": "limitedUploadDocuments",
     "CaseFieldID": "pageShow",
-    "PageFieldDisplayOrder": 1,
+    "PageFieldDisplayOrder": 2,
     "DisplayContext": "READONLY",
     "PageID": "routing",
     "PageDisplayOrder": 1,
     "PageColumnNumber": 1,
-    "FieldShowCondition": "uploadDocumentsRouter = \"DO_NOT_SHOW\""
+    "FieldShowCondition": "manageDocumentsAction = \"DO_NOT_SHOW\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -34,7 +34,7 @@
     "PageID": "new",
     "PageDisplayOrder": 2,
     "PageColumnNumber": 1,
-    "PageShowCondition": "uploadDocumentsRouter = \"UPLOAD\" OR pageShow = \"No\""
+    "PageShowCondition": "manageDocumentsAction = \"UPLOAD\" OR pageShow = \"No\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -57,9 +57,9 @@
     "PageID": "select",
     "PageDisplayOrder": 3,
     "PageColumnNumber": 1,
-    "PageShowCondition": "uploadDocumentsRouter = \"AMEND\" OR uploadDocumentsRouter = \"DELETE\"",
+    "PageShowCondition": "manageDocumentsAction = \"AMEND\" OR manageDocumentsAction = \"DELETE\"",
     "CallBackURLMidEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/manage-docs/get-doc/mid-event",
-    "FieldShowCondition": "uploadDocumentsRouter = \"AMEND\""
+    "FieldShowCondition": "manageDocumentsAction = \"AMEND\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -71,7 +71,7 @@
     "PageID": "select",
     "PageDisplayOrder": 3,
     "PageColumnNumber": 1,
-    "FieldShowCondition": "uploadDocumentsRouter = \"DELETE\""
+    "FieldShowCondition": "manageDocumentsAction = \"DELETE\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -94,7 +94,7 @@
     "PageID": "amend",
     "PageDisplayOrder": 4,
     "PageColumnNumber": 1,
-    "PageShowCondition": "uploadDocumentsRouter = \"AMEND\""
+    "PageShowCondition": "manageDocumentsAction = \"AMEND\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -139,7 +139,7 @@
     "PageID": "delete",
     "PageDisplayOrder": 5,
     "PageColumnNumber": 1,
-    "PageShowCondition": "uploadDocumentsRouter = \"DELETE\"",
+    "PageShowCondition": "manageDocumentsAction = \"DELETE\"",
     "Comment": "Complex to overwrite the file field label"
   }
 ]
