+++ ccd-definition/CaseEventToFields/CareSupervision/uploadStandardDirections.json
@@ -3,9 +3,9 @@
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseEventID": "uploadStandardDirections",
-    "CaseFieldID": "uploadDocumentsRouter",
+    "CaseFieldID": "manageDocumentsAction",
     "CaseEventFieldHint": "To replace the standard directions, choose 'Upload a new case document'",
-    "PageFieldDisplayOrder": 2,
+    "PageFieldDisplayOrder": 1,
     "DisplayContext": "MANDATORY",
     "PageID": "routing",
     "PageDisplayOrder": 1,
@@ -18,12 +18,12 @@
     "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseEventID": "uploadStandardDirections",
     "CaseFieldID": "pageShow",
-    "PageFieldDisplayOrder": 1,
+    "PageFieldDisplayOrder": 2,
     "DisplayContext": "READONLY",
     "PageID": "routing",
     "PageDisplayOrder": 1,
     "PageColumnNumber": 1,
-    "FieldShowCondition": "uploadDocumentsRouter = \"DO_NOT_SHOW\""
+    "FieldShowCondition": "manageDocumentsAction = \"DO_NOT_SHOW\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -35,7 +35,7 @@
     "PageID": "new",
     "PageDisplayOrder": 2,
     "PageColumnNumber": 1,
-    "PageShowCondition": "uploadDocumentsRouter = \"UPLOAD\" OR pageShow = \"No\""
+    "PageShowCondition": "manageDocumentsAction = \"UPLOAD\" OR pageShow = \"No\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -91,9 +91,9 @@
     "PageID": "select",
     "PageDisplayOrder": 3,
     "PageColumnNumber": 1,
-    "PageShowCondition": "uploadDocumentsRouter = \"AMEND\" OR uploadDocumentsRouter = \"DELETE\"",
+    "PageShowCondition": "manageDocumentsAction = \"AMEND\" OR manageDocumentsAction = \"DELETE\"",
     "CallBackURLMidEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/manage-docs/get-doc/mid-event",
-    "FieldShowCondition": "uploadDocumentsRouter = \"AMEND\""
+    "FieldShowCondition": "manageDocumentsAction = \"AMEND\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -105,7 +105,7 @@
     "PageID": "select",
     "PageDisplayOrder": 3,
     "PageColumnNumber": 1,
-    "FieldShowCondition": "uploadDocumentsRouter = \"DELETE\""
+    "FieldShowCondition": "manageDocumentsAction = \"DELETE\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -128,7 +128,7 @@
     "PageID": "amend",
     "PageDisplayOrder": 4,
     "PageColumnNumber": 1,
-    "PageShowCondition": "uploadDocumentsRouter = \"AMEND\""
+    "PageShowCondition": "manageDocumentsAction = \"AMEND\""
   },
   {
     "LiveFrom": "01/01/2017",
@@ -173,7 +173,7 @@
     "PageID": "delete",
     "PageDisplayOrder": 5,
     "PageColumnNumber": 1,
-    "PageShowCondition": "uploadDocumentsRouter = \"DELETE\"",
+    "PageShowCondition": "manageDocumentsAction = \"DELETE\"",
     "Comment": "Complex to overwrite the file field label"
   }
 ]
