+++ service/src/main/java/uk/gov/hmcts/reform/fpl/controllers/ManageDocumentsController.java
@@ -92,7 +92,7 @@ public class ManageDocumentsController {
             "editedCourtDocument",
             "deletedCourtDocument",
             "courtDocumentList",
-            "uploadDocumentsRouter",
+            "manageDocumentsAction",
             "originalCourtDocument",
             "pageShow"
         );
