+++ service/src/main/java/uk/gov/hmcts/reform/fpl/enums/ccd/fixedlists/ManageDocumentsAction.java
@@ -1,6 +1,6 @@
 package uk.gov.hmcts.reform.fpl.enums.ccd.fixedlists;
 
-public enum DocumentRouter {
+public enum ManageDocumentsAction {
     UPLOAD,
     AMEND,
     DELETE
