+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/CaseData.java
@@ -13,7 +13,7 @@ import uk.gov.hmcts.reform.fpl.enums.CaseExtensionTime;
 import uk.gov.hmcts.reform.fpl.enums.EPOType;
 import uk.gov.hmcts.reform.fpl.enums.OrderStatus;
 import uk.gov.hmcts.reform.fpl.enums.State;
-import uk.gov.hmcts.reform.fpl.enums.ccd.fixedlists.DocumentRouter;
+import uk.gov.hmcts.reform.fpl.enums.ccd.fixedlists.ManageDocumentsAction;
 import uk.gov.hmcts.reform.fpl.exceptions.NoHearingBookingException;
 import uk.gov.hmcts.reform.fpl.model.common.C2DocumentBundle;
 import uk.gov.hmcts.reform.fpl.model.common.Document;
@@ -643,7 +643,7 @@ public class CaseData {
         return defaultIfNull(limitedCourtAdminDocuments, new ArrayList<>());
     }
 
-    private final DocumentRouter uploadDocumentsRouter;
+    private final ManageDocumentsAction manageDocumentsAction;
     private final CourtAdminDocument editedCourtDocument;
     private final Object courtDocumentList;
 }
